Made the memory allocation code a little more stable. Also made a start on the memory map interpretation.

+++ andromeda/include/types.h
@@ -22,7 +22,7 @@ typedef unsigned char boolean;
 #define FALSE 0
 #define TRUE !FALSE
 
-typedef unsigned int size_t;
+typedef unsigned long size_t;
 
 typedef struct
 {
